Make stringifying of a message conditional
Do not stringify the message if it's already a string.

Rationale:

If a user passes a string message to publish, it's probably her
intention to send it's raw content to the subscribers (e.g. via e-mail).
Currently such a message would be passed to JSON.stringify first and the
result would be a JSONO representation of a string (i.e. surrounded with
quotes: '"message here"'). Subsequently e-mail notifications would also
present these quotes, which looks weird.

Note:

Note that this is a breaking change - some users may depend on the
serialization of raw strings. Thus if this PR is to be merged, I would
advise a major version bump.

To achieve previous behavior a user would have to stringify the message
herself before passing it to publish:

    SNS.publish(topic, JSON.stringify(message));